% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueConversion.R
\name{pvalueConversion}
\alias{pvalueConversion}
\title{pvalueCoversion}
\usage{
pvalueConversion(x, pvalueBase = 1L, ...)
}
\arguments{
\item{x}{GRanges objects set of peak Interval}

\item{pvalueBase}{parameter for user defined pvalue format ( - log(p.value), -10 log(p.value), -100 log(p.value))}
}
\value{
GRanges
}
\description{
MSPC Project - Bioconductor Package for Multiple Sample Peak Calling
}
\details{
In standard BED file, significant value of peak signal defined as Score column,
so we need to convert it as p.value (- log(p.value), -10 log(p.value), -100 log(p.value))
}
\author{
Julaiti Shayiding
.pvalueConversion(gr, 1L)
}

